requirements.txt
wechatpy>=1.8.15
requests>=2.31.0
python-dotenv>=1.0.0
Flask>=2.3.3
cryptography>=41.0.0
PyJWT>=2.8.0